requirements/base.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
phonenumbers==8.13.27
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.29.0
social-auth-app-django==5.4.3
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
celery==5.3.4
Pillow


